[frontend]sound の型定義を別ファイルに切り出した
