+++ frontend/src/types/sound.ts
@@ -1,0 +1,10 @@
+export type sound = {
+  id: number;
+  isrc: string;
+  thumbnailUrl: string;
+  title: string;
+  artist: string;
+  spotifyUrl: string;
+  appleMusicUrl: string;
+  kkboxUrl: string;
+};
