--- frontend/src/types/vue.d.ts
@@ -1,22 +1,0 @@
-// vuex.d.ts
-// import { ComponentCustomProperties } from 'vue';
-// import { Store } from 'vuex';
-// import { result } from '@/types/result';
-//
-// declare module '@vue/runtime-core' {
-//   // declare your own store states
-//   interface State {
-//     keyword: string;
-//     results: result[];
-//     isLoading: boolean;
-//     currentPage: number;
-//     showError: boolean;
-//   }
-//
-//   // provide typings for `this.$store`
-//   declare module '@vue/runtime-core' {
-//     interface ComponentCustomProperties {
-//       $store: Store<State>
-//     }
-//   }
-// }
